BackEnd/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.4
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.7.4.3
